Tooltip now base class, hints wont render without content
