+++ src/components/Tooltip/index.js
@@ -1,10 +1,9 @@
 // @flow
 
 import React from 'react'
+import ReactDOM from 'react-dom'
 import { Portal } from 'react-portal'
 import { GlobalStylesScope } from 'components/ThemeProvider'
-import Hint from 'blocks/Hints'
-import type { Coordinates } from 'blocks/Hints'
 import TextSmall from 'components/Typography/TextSmall'
 import {
   TooltipBackground,
@@ -14,15 +13,24 @@ import {
   PositionWrapper,
   PlacementWrapper,
 } from './styled'
+/* eslint-disable react/no-unused-prop-types */
+
+export type Coordinates = {
+  left: number,
+  top: number,
+  width: number,
+  height: number,
+}
+
+type GetCoordinates = (node : Element) => Coordinates
 
 type PortalProps = {
   isOpen: boolean,
   coords: Coordinates | null,
   placement: string,
-  content: string | Element,
+  content: string | Element | any | null,
   success: ?boolean,
   error: ?boolean,
-  shouldRender: boolean,
 }
 
 const TooltipPortal = (props : PortalProps) => {
@@ -33,10 +41,9 @@ const TooltipPortal = (props : PortalProps) => {
     content,
     success,
     error,
-    shouldRender,
   } = props
 
-  return ((shouldRender && isOpen && coords)
+  return ((content && isOpen && coords)
     ? <Portal>
       <GlobalStylesScope>
         <TooltipContainer
@@ -77,11 +84,13 @@ const TooltipPortal = (props : PortalProps) => {
 
 type TooltipProps = {
   children: Object | Element,
-  content: string | Element,
+  content: string | Element| any | null,
   placement: string,
   success: ?boolean,
   error: ?boolean,
-  shouldRender: boolean,
+  dotCentering: ?boolean,
+  header: ?string,
+  align: ?string,
 }
 
 type TooltipState = {
@@ -89,15 +98,60 @@ type TooltipState = {
 }
 
 /* eslint-disable react/prop-types */
-class Tooltip extends Hint {
-  props: TooltipProps
-  state: TooltipState
+class Tooltip extends React.Component<TooltipProps, TooltipState> {
   static defaultProps = {
     placement: 'bottom',
     success: false,
     error: false,
-    shouldRender: true,
+    dotCentering: false,
   }
+
+  state = {
+    isOpen: false,
+  }
+
+  componentDidMount() {
+    if (this.childRef !== null) {
+      this.coords = this.getCoordinates(this.childRef)
+    }
+  }
+
+  componentWillUnmount() {
+    clearTimeout(this.hoverTimeout)
+  }
+  /* eslint-disable react/no-find-dom-node */
+  getCoordinates = (node) : GetCoordinates => {
+    const availableNode = ReactDOM.findDOMNode(node)
+    if (availableNode) {
+      const rect = ReactDOM.findDOMNode(node).getBoundingClientRect()
+      return {
+        width: rect.width,
+        height: rect.height,
+        left: rect.left + window.pageXOffset,
+        top: rect.top + window.pageYOffset,
+      }
+    }
+  }
+
+  handleMouseOut = () => {
+    clearTimeout(this.hoverTimeout)
+    this.setState({
+      isOpen: false,
+    })
+  }
+
+  handleMouseOver = () => {
+    this.hoverTimeout = setTimeout(() => {
+      this.coords = this.getCoordinates(this.childRef)
+      this.setState({
+        isOpen: true,
+      })
+    }, 150)
+  }
+
+  childRef = null
+  coords = null
+  hoverTimeout = null
 
   render() {
     const {
@@ -134,7 +188,7 @@ class Tooltip extends Hint {
 }
 
 // Proxy for possibility to transfer ref to any children
-class TooltipChildrenProxy extends React.Component {
+class TooltipChildrenProxy extends React.Component <void, void> {
   render() {
     const { children, ...props } = this.props
     return React.cloneElement(React.Children.only(children), props)
